feat(add-alarm): Improve radius based selection, add inScaleMode banner
