+++ lib/l10n/app_en.arb
@@ -569,5 +569,8 @@
   "locationAlarm_notification_description": "Tap for more information",
   "androidNotificationChannel_locationAlarms_name": "Location Alarms",
   "androidNotificationChannel_locationAlarms_description": "Receive notifications for location alarms",
-  "location_addAlarm_title": "Add Alarm"
+  "location_addAlarm_title": "Add Alarm",
+  "location_addAlarm_radiusBased_title": "Add radius-based Alarm",
+  "location_addAlarm_radiusBased_isInScaleMode": "You can now scale the radius by pinching and zooming into the map. Tap on the screen to go back to the normal mode.",
+  "location_addAlarm_radiusBased_addLabel": "Select region"
 }
